# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type FinanceTransactionDirection = typeof FinanceTransactionDirection[keyof typeof FinanceTransactionDirection];


export const FinanceTransactionDirection = {
  in: 'in',
  out: 'out',
} as const;

export interface FinanceTransaction {
  id: string;
  title: string;
  amount: number;
  direction: FinanceTransactionDirection;
  createdAt: string;
}

export interface FinanceDashboard {
  balance: number;
  currency: string;
  transactions: FinanceTransaction[];
}

export interface FinanceCommandRequest {
  text: string;
}

export type FinanceCommandResultIntent = typeof FinanceCommandResultIntent[keyof typeof FinanceCommandResultIntent];


export const FinanceCommandResultIntent = {
  send_money: 'send_money',
  check_balance: 'check_balance',
  apply_loan: 'apply_loan',
} as const;

export interface FinanceCommandResult {
  intent: FinanceCommandResultIntent;
  message: string;
  balance?: number | null;
  transaction?: FinanceTransaction | null;
}

export interface FinanceTranscriptionRequest {
  /** Base64-encoded microphone recording */
  audioBase64: string;
  /** MIME type of the recording */
  mimeType?: string;
}

export interface FinanceTranscriptionResult {
  text: string;
}