# lib/api-zod/src/generated/types/financeCommandResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FinanceCommandResultIntent } from './financeCommandResultIntent';
import type { FinanceTransaction } from './financeTransaction';

export interface FinanceCommandResult {
  intent: FinanceCommandResultIntent;
  message: string;
  balance?: number | null;
  transaction?: FinanceTransaction | null;
}