# lib/api-zod/src/generated/types/financeDashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FinanceTransaction } from './financeTransaction';

export interface FinanceDashboard {
  balance: number;
  currency: string;
  transactions: FinanceTransaction[];
}