# lib/api-zod/src/generated/types/financeTransaction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FinanceTransactionDirection } from './financeTransactionDirection';

export interface FinanceTransaction {
  id: string;
  title: string;
  amount: number;
  direction: FinanceTransactionDirection;
  createdAt: string;
}