# lib/api-zod/src/generated/types/financeTranscriptionRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FinanceTranscriptionRequest {
  /** Base64-encoded microphone recording */
  audioBase64: string;
  /** MIME type of the recording */
  mimeType?: string;
}